+++ docs/change-log.md
@@ -9,12 +9,17 @@
 * Multiple <class> within <entity> needs some work
 * Approach to define generator policy needs to be discussed
 
-## 08 April 2014: V 1.1
+## 07 April 2014: V 1.1
 
 * Major refactor for code legibility, licensing, and javadoc
 * The *property* no longer contains *class* but instead itself holds the qualified name
 * The tag *property* has been replaced with *relationship*.
 * The *class* element is now *type*.
+* The *value_generator* is now *instance_generator*
+* Multiple *type* possible within entity, variables hold lists now
+* X3ML root tag now has *language*, which becomes default instance generator language
+* Labels generator arguments now have optional *language*
+* Argument types and source_type now lowercase *xpath* or *constant*
 
 ---